backend/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
pymongo==3.12.0
djongo==1.3.6
django-cors-headers==3.14.0